Fix access logs and visualization
Access logs were missing a lot of access denied, which are now added and the UI was displaying only the role of the person trying the access, not the actual result.

So now we are logging the card, the role and the result of the opening.

For example a standard card could have failed because it has expired, it's disabled or it's outside of the opening hours. In all these cases we should add a line in the access logs saying that the person with the specific role had a denied access.